litesata: use (some) settings from vivado 2015.1, try to follow all ug476 recommendations to initialize GTX (...), remove automatic reset on top.
Works fine @ 3Gbps, still not working @6.0Gbps

+++ misoclib/mem/litesata/example_designs/test/bist.py
@@ -191,7 +191,7 @@ if __name__ == "__main__":
 
                 print("sector={:d}({:d}MB) wr_speed={:4.2f}MB/s rd_speed={:4.2f}MB/s errors={:d} retry={:d}".format(
                     sector,
-                    run_sectors*logical_sector_size/MB,
+                    int(run_sectors*logical_sector_size/MB),
                     write_speed/MB,
                     read_speed/MB,
                     write_errors + read_errors,
